requirements.txt
# Core dependencies for ContentCreationEngine

# Environment and Configuration
python-dotenv>=1.0.0

# AI Providers
openai>=1.0.0
httpx>=0.25.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
praw>=7.7.0  # Reddit API
google-api-python-client>=2.100.0  # YouTube Data API v3

# Scheduling
APScheduler>=3.10.0

# Data Processing
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
